pyproject.toml
[project]
name = "edcraft-backend"
version = "0.1.0"
description = "EdCraft FastAPI Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.17.2",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.31.0",
    "authlib>=1.3.0",
    "edcraft-engine",
    "fastapi[standard]>=0.125.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.7.0",
    "pymustache>=0.3",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.45",
    "input-gen",
]

[tool.uv.sources]
edcraft-engine = { git = "https://github.com/edcraft-org/edcraft-engine.git" }
input-gen = { git = "https://github.com/edcraft-org/input-gen.git" }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "types-authlib>=1.6.6.20260130",
    "types-python-jose>=3.5.0.20250531",
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "integration: Integration tests (HTTP endpoints)",
    "auth: Authentication and authorization router tests",
    "users: User router tests",
    "folders: Folder router tests",
    "questions: Question router tests",
    "question_banks: Question bank router tests",
    "assessments: Assessment router tests",
    "question_templates: Question template router tests",
    "question_template_banks: Question template bank router tests",
    "assessment_templates: Assessment template router tests",
    "question_generation: Question generation router tests",
    "input_generator: Input generator router tests",
    "collaborators: Collaborator router tests",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "UP", "B", "A", "S", "T20", "Q"]
ignore = [
    "B008",  # Allow function calls in default arguments (FastAPI Depends pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # Allow asserts in tests
    "S106", # Allow passwords in tests
]
